mirage: Split test files into one file per endpoint
Having files with 600+ lines that test 7 different endpoints is starting to become unmaintainable...
