+++ tests/mirage/confirm/put-by-id-test.js
@@ -1,0 +1,53 @@
+import { module, test } from 'qunit';
+
+import fetch from 'fetch';
+
+import { setupTest } from '../../helpers';
+import setupMirage from '../../helpers/setup-mirage';
+
+module('Mirage | /me', function (hooks) {
+  setupTest(hooks);
+  setupMirage(hooks);
+
+  module('PUT /api/v1/confirm/:token', function () {
+    test('returns `ok: true` for a known token (unauthenticated)', async function (assert) {
+      let user = this.server.create('user', { emailVerificationToken: 'foo' });
+      assert.strictEqual(user.emailVerified, false);
+
+      let response = await fetch('/api/v1/confirm/foo', { method: 'PUT' });
+      assert.equal(response.status, 200);
+
+      let responsePayload = await response.json();
+      assert.deepEqual(responsePayload, { ok: true });
+
+      user.reload();
+      assert.strictEqual(user.emailVerified, true);
+    });
+
+    test('returns `ok: true` for a known token (authenticated)', async function (assert) {
+      let user = this.server.create('user', { emailVerificationToken: 'foo' });
+      assert.strictEqual(user.emailVerified, false);
+
+      this.server.create('mirage-session', { user });
+
+      let response = await fetch('/api/v1/confirm/foo', { method: 'PUT' });
+      assert.equal(response.status, 200);
+
+      let responsePayload = await response.json();
+      assert.deepEqual(responsePayload, { ok: true });
+
+      user.reload();
+      assert.strictEqual(user.emailVerified, true);
+    });
+
+    test('returns an error for unknown tokens', async function (assert) {
+      let response = await fetch('/api/v1/confirm/unknown', { method: 'PUT' });
+      assert.equal(response.status, 400);
+
+      let responsePayload = await response.json();
+      assert.deepEqual(responsePayload, {
+        errors: [{ detail: 'Email belonging to token not found.' }],
+      });
+    });
+  });
+});
