+++ tests/mirage/crates/follow/get-test.js
@@ -1,0 +1,60 @@
+import { module, test } from 'qunit';
+
+import fetch from 'fetch';
+
+import { setupTest } from '../../../helpers';
+import setupMirage from '../../../helpers/setup-mirage';
+
+module('Mirage | Crates', function (hooks) {
+  setupTest(hooks);
+  setupMirage(hooks);
+
+  module('GET /api/v1/crates/:crateId/following', function () {
+    test('returns 403 if unauthenticated', async function (assert) {
+      let response = await fetch('/api/v1/crates/foo/following');
+      assert.equal(response.status, 403);
+
+      let responsePayload = await response.json();
+      assert.deepEqual(responsePayload, {
+        errors: [{ detail: 'must be logged in to perform that action' }],
+      });
+    });
+
+    test('returns 404 for unknown crates', async function (assert) {
+      let user = this.server.create('user');
+      this.authenticateAs(user);
+
+      let response = await fetch('/api/v1/crates/foo/following');
+      assert.equal(response.status, 404);
+
+      let responsePayload = await response.json();
+      assert.deepEqual(responsePayload, { errors: [{ detail: 'Not Found' }] });
+    });
+
+    test('returns true if the authenticated user follows the crate', async function (assert) {
+      let crate = this.server.create('crate', { name: 'rand' });
+
+      let user = this.server.create('user', { followedCrates: [crate] });
+      this.authenticateAs(user);
+
+      let response = await fetch('/api/v1/crates/rand/following');
+      assert.equal(response.status, 200);
+
+      let responsePayload = await response.json();
+      assert.deepEqual(responsePayload, { following: true });
+    });
+
+    test('returns false if the authenticated user is not following the crate', async function (assert) {
+      this.server.create('crate', { name: 'rand' });
+
+      let user = this.server.create('user');
+      this.authenticateAs(user);
+
+      let response = await fetch('/api/v1/crates/rand/following');
+      assert.equal(response.status, 200);
+
+      let responsePayload = await response.json();
+      assert.deepEqual(responsePayload, { following: false });
+    });
+  });
+});
