+++ tests/mirage/crates/owner-team-test.js
@@ -1,0 +1,56 @@
+import { module, test } from 'qunit';
+
+import fetch from 'fetch';
+
+import { setupTest } from '../../helpers';
+import setupMirage from '../../helpers/setup-mirage';
+
+module('Mirage | Crates', function (hooks) {
+  setupTest(hooks);
+  setupMirage(hooks);
+
+  module('GET /api/v1/crates/:id/owner_team', function () {
+    test('returns 404 for unknown crates', async function (assert) {
+      let response = await fetch('/api/v1/crates/foo/owner_team');
+      assert.equal(response.status, 404);
+
+      let responsePayload = await response.json();
+      assert.deepEqual(responsePayload, { errors: [{ detail: 'Not Found' }] });
+    });
+
+    test('empty case', async function (assert) {
+      this.server.create('crate', { name: 'rand' });
+
+      let response = await fetch('/api/v1/crates/rand/owner_team');
+      assert.equal(response.status, 200);
+
+      let responsePayload = await response.json();
+      assert.deepEqual(responsePayload, {
+        teams: [],
+      });
+    });
+
+    test('returns the list of teams that own the specified crate', async function (assert) {
+      let team = this.server.create('team', { name: 'maintainers' });
+      let crate = this.server.create('crate', { name: 'rand' });
+      this.server.create('crate-ownership', { crate, team });
+
+      let response = await fetch('/api/v1/crates/rand/owner_team');
+      assert.equal(response.status, 200);
+
+      let responsePayload = await response.json();
+      assert.deepEqual(responsePayload, {
+        teams: [
+          {
+            id: 1,
+            avatar: 'https://avatars1.githubusercontent.com/u/14631425?v=4',
+            kind: 'team',
+            login: 'github:rust-lang:maintainers',
+            name: 'maintainers',
+            url: 'https://github.com/rust-lang',
+          },
+        ],
+      });
+    });
+  });
+});
