+++ tests/mirage/me/tokens/create-test.js
@@ -1,0 +1,51 @@
+import { module, test } from 'qunit';
+
+import fetch from 'fetch';
+import timekeeper from 'timekeeper';
+
+import { setupTest } from '../../../helpers';
+import setupMirage from '../../../helpers/setup-mirage';
+
+module('Mirage | /me', function (hooks) {
+  setupTest(hooks);
+  setupMirage(hooks);
+
+  module('PUT /api/v1/me/tokens', function () {
+    test('creates a new API token', async function (assert) {
+      timekeeper.freeze(new Date('2017-11-20T11:23:45Z'));
+
+      let user = this.server.create('user');
+      this.server.create('mirage-session', { user });
+
+      let body = JSON.stringify({ api_token: { name: 'foooo' } });
+      let response = await fetch('/api/v1/me/tokens', { method: 'PUT', body });
+      assert.equal(response.status, 200);
+
+      let token = this.server.schema.apiTokens.all().models[0];
+      assert.ok(token);
+
+      let responsePayload = await response.json();
+      assert.deepEqual(responsePayload, {
+        api_token: {
+          id: 1,
+          created_at: '2017-11-20T11:23:45.000Z',
+          last_used_at: null,
+          name: 'foooo',
+          revoked: false,
+          token: token.token,
+        },
+      });
+    });
+
+    test('returns an error if unauthenticated', async function (assert) {
+      let body = JSON.stringify({ api_token: {} });
+      let response = await fetch('/api/v1/me/tokens', { method: 'PUT', body });
+      assert.equal(response.status, 403);
+
+      let responsePayload = await response.json();
+      assert.deepEqual(responsePayload, {
+        errors: [{ detail: 'must be logged in to perform that action' }],
+      });
+    });
+  });
+});
