+++ tests/mirage/me/tokens/delete-by-id-test.js
@@ -1,0 +1,42 @@
+import { module, test } from 'qunit';
+
+import fetch from 'fetch';
+
+import { setupTest } from '../../../helpers';
+import setupMirage from '../../../helpers/setup-mirage';
+
+module('Mirage | /me', function (hooks) {
+  setupTest(hooks);
+  setupMirage(hooks);
+
+  module('DELETE /api/v1/me/tokens/:tokenId', function () {
+    test('revokes an API token', async function (assert) {
+      let user = this.server.create('user');
+      this.server.create('mirage-session', { user });
+
+      let token = this.server.create('api-token', { user });
+
+      let response = await fetch(`/api/v1/me/tokens/${token.id}`, { method: 'DELETE' });
+      assert.equal(response.status, 200);
+
+      let responsePayload = await response.json();
+      assert.deepEqual(responsePayload, {});
+
+      let tokens = this.server.schema.apiTokens.all().models;
+      assert.equal(tokens.length, 0);
+    });
+
+    test('returns an error if unauthenticated', async function (assert) {
+      let user = this.server.create('user');
+      let token = this.server.create('api-token', { user });
+
+      let response = await fetch(`/api/v1/me/tokens/${token.id}`, { method: 'DELETE' });
+      assert.equal(response.status, 403);
+
+      let responsePayload = await response.json();
+      assert.deepEqual(responsePayload, {
+        errors: [{ detail: 'must be logged in to perform that action' }],
+      });
+    });
+  });
+});
